requirements.txt
flask
flask-restful
python-dotenv
flask_sqlalchemy
flask_wtf
wtforms